Pull out a VisitForm component from book-a-visit
